requirements.txt
streamlit
fastapi
uvicorn
python-multipart
requests
pymupdf
rapidocr_onnxruntime
opencv-python
pillow
docx2pdf